feat: 유저권한 여부에 따라 게시글 삭제하도록 기능 추가

+++ src/main/java/org/example/toy_restboard/service/BoardService.java
@@ -4,10 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.example.toy_restboard.domain.dto.boarddto.BoardCreateDto;
 import org.example.toy_restboard.domain.dto.boarddto.BoardDto;
 import org.example.toy_restboard.domain.dto.boarddto.BoardUpdateDto;
+import org.example.toy_restboard.domain.dto.userdto.AccountContext;
 import org.example.toy_restboard.domain.entity.Board;
 import org.example.toy_restboard.domain.entity.User;
 import org.example.toy_restboard.repository.BoardRepository;
 import org.example.toy_restboard.repository.UserRepository;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,9 +23,11 @@ public class BoardService {
     private final BoardRepository boardRepository;
     private final UserRepository userRepository;
 
+
     @Transactional
-    public void createBoard(BoardCreateDto boardCreateDto) {
-        User user = userRepository.findByLoginId(boardCreateDto.getUserId());
+    public void createBoard(BoardCreateDto boardCreateDto, AccountContext accountContext) {
+        User user = userRepository.findByLoginId(accountContext.getUsername()).orElseThrow();
+
         Board board = new Board();
         board.setUser(user);
         board.setTitle(boardCreateDto.getTitle());
@@ -32,9 +36,12 @@ public class BoardService {
     }
 
     @Transactional
-    public void updateBoard(Long boardId, BoardUpdateDto boardUpdateDto) {
+    public void updateBoard(Long boardId, BoardUpdateDto boardUpdateDto, AccountContext accountContext) {
         Optional<Board> optionalBoard = boardRepository.findById(boardId);
         Board board = optionalBoard.orElse(null);
+        if (!board.getUser().getLoginId().equals(accountContext.getUsername())) {
+            throw new RuntimeException("게시글을 수정할 권한이 없습니다.");
+        }
         board.setTitle(boardUpdateDto.getTitle());
         board.setContent(boardUpdateDto.getContent());
         boardRepository.save(board);
@@ -55,11 +62,16 @@ public class BoardService {
     }
 
     @Transactional
-    public void deleteBoard(Long boardId) {
+    public void deleteBoard(Long boardId, @AuthenticationPrincipal AccountContext accountContext) {
         Optional<Board> optionalBoard = boardRepository.findById(boardId);
         Board board = optionalBoard.get();
+        if (!board.getUser().getLoginId().equals(accountContext.getUsername())) {
+            throw new RuntimeException("게시글을 삭제할 권한이 없습니다.");
+        }
         boardRepository.delete(board);
+
     }
+
     private BoardDto mapToDto(Board board) {
         return new BoardDto(
                 board.getBoardId(),
